+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added general platform that identifies the plugin with the API.
 - Changed name of ReadOnly field to not occupy reserved named.
 - Updated dependencies.
 - Fixed token storage client id issue.