+++ CHANGELOG.md
@@ -4,7 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added general platform that identifies the pluin with the API.
+- Added general platform that identifies the plugin with the API.
+- Changed name of ReadOnly field to not occupy reserved named.
+- Updated dependencies.
+- Fixed token storage client id issue.
+- Resolved race condition issue in return/callback.
 
 ## [1.0.3] - 2023-08-23
 - Implemented payment creation through the API, instead of posting to payment window.