+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.6] - 2026-08-04
 - Fix TypeError when throwing ValidatorException with string message
 - Added unit tests for capture, refund, cancel and void
 - Validate HMAC and session on payment decline